pi_cond: Add cancellation point for condition waits
Per POSIX "A condition wait (whether timed or not) is a cancellation
point. When the cancelability enable state of a thread is set to
PTHREAD_CANCEL_DEFERRED, a side effect of acting upon a cancellation
request while in a condition wait is that the mutex is (in effect)
re-acquired before calling the first cancellation cleanup handler. The
effect is as if the thread were unblocked, allowed to execute up to
the point of returning from the call to pthread_cond_wait() or
pthread_cond_timedwait(), but at that point notices the cancellation
request and instead of returning to the caller of pthread_cond_wait()
or pthread_cond_timedwait(), starts the thread cancellation
activities, which includes calling cancellation cleanup handlers.

A thread that has been unblocked because it has been canceled while
blocked in a call to pthread_cond_wait() or pthread_cond_timedwait()
does not consume any condition signal that may be directed
concurrently at the condition variable if there are other threads
blocked on the condition variable."

We don't have a good reason to deviate from POSIX here so implement
cancellation for pi_cond_wait() and pi_cond_timedwait() by turning on
asynchronous cancellation before we block on the futex call and
installing a cleanup handler which re-acquires the user mutex as
needed before starting the rest of the thread cancellation activities.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>