pi_cond: Remove the 'nothing woke up case' from pi_cond_signal
If no threads were woken by the futex_cmp_requeue_pi() call it just
means the kernel had nobody to wake up. We shouldn't try again because
per POSIX:

  "The pthread_cond_broadcast() and pthread_cond_signal() functions
  shall have no effect if there are no threads currently blocked on
  cond."

Remove the check for ret == 0 and the associated code from
pi_cond_signal(). This also makes it easier to later merge the code
with the common code in pi_cond_broadcast() since the later doesn't
have special handling for this case.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>