pi_cond: Account for pending waiters in pi_cond_destroy
Per POSIX:

  "It shall be safe to destroy an initialized condition variable upon
  which no threads are currently blocked. Attempting to destroy a
  condition variable upon which other threads are currently blocked
  results in undefined behavior."

  "If an implementation detects that the value specified by the cond
  argument to pthread_cond_destroy() or pthread_cond_init() refers to
  a condition variable that is in use (for example, in a
  pthread_cond_wait() call) by another thread, or detects that the
  value specified by the cond argument to pthread_cond_init() refers
  to an already initialized condition variable, it is recommended that
  the function should fail and report an [EBUSY] error."

Implement this logic by relying on the fact that the
futex_wait_requeue_pi() call in pi_cond_signal_common() returns the
number of threads woken. pi_cond_destroy() first issues a broadcast
signal to any waiters that might be waiting on the conditional
variable. If no threads were woken up (ret == 0) it is safe to destroy
the conditional variable. Otherwise if we got a positive return value
it means there were waiters and we return EBUSY. If we got another
type of error we return its positive equivalent value.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>