pi_cond: Remove private mutex and rely instead on user mutex
Having to deal with a private lock for accessing the internal state of
a conditional variable is a source of trouble because it is hard to
correctly deal with errors coming from this private mutex.

The librtpi API creates a strong association between the conditional
variable and the user mutex via pi_cond_init(). This allows us to
eliminate the private lock and instead use the user mutex for
protecting access to internal condvar fields.

Per POSIX cond_wait() and cond_timedwait() already have to be called
and return with the mutex held. We can use this to our advantage when
modifying the internal condvar state in pi_cond_wait/timedwait.

Additionally, if not already held, acquire the user mutex in
pi_cond_signal_common() when modifying the internal state. This
imposes ordering between the waiters and signalers however this is
desirable for real-time applications and it is recommended by POSIX
which states:

  "if predictable scheduling behavior is required, then that mutex
  shall be locked by the thread calling pthread_cond_broadcast() or
  pthread_cond_signal()."

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>