pi_cond: Eliminate the pending_wait counter
The 'pending_wait' counter is used to skip the futex_cmp_requeue_pi()
call if there are no waiters in the pi_cond_signal* paths. This adds a
lot of complexity to the error paths in pi_cond_timedwait().

It is simpler to let the kernel handle the 'no waiters' case for us in
a race free manner. This simplifies the code and eliminates the need
to grab the private lock to manipulate this counter.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>