pi_cond: Make error codes consistent with glibc counterparts
The current implementation for pi_cond* functions returns 0 on success
and a negative error number to indicate the error.

The glibc implementation for pthread_cond* returns 0 on success and a
positive error number on failure (e.g. EINVAL instead of -EINVAL for
an invalid parameter).

In order to avoid porting/programming mistakes, match the glibc
behavior by changing the return paths in pi_cond to return a positive
error number on failure.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>