pi_cond: Fix user mutex handling in pi_cond_timedwait
The current implementation of pi_cond_timedwait() improperly tries to
acquire the user mutex in the normal wake-up case where the kernel
already acquired the mutex for us. It also doesn't acquire the mutex
for the error case where it should. On error the kernel does not lock
the mutex for us but pi_cond_timedwait() is expected to always return
with the user mutex locked. Another bug is the fact that for the
normal wake-up case it unlocks the 'cond->priv_mut' private mutex w/o
first locking it before accessing the private 'pending_wait/wake'
counters.

For the normal wake-up case remove the double locking, and add
'cond->priv_mut' locking before accessing the 'pending_wake/wait'
counters.

For the return with error case grab the user mutex 'cond->mutex'
before returning.

Signed-off-by: Gratian Crisan <gratian.crisan@ni.com>